Create res_rtp_asterisk.c.diff
This adds the extra test to locate the character 'F' or 'f' in the SendDTMF dialplan application and convert it to a SIP INFO Event 16 (Hook flash, deprecated but remains reserved in RFC2833 and RFC4733).  The SIP endpoint must support using dtmfmode=rfc2833 or dtmfmode=rfc4733 but it will work with either chan_sip or chan_pjsip.  This edit must be combined with edits in main/app.c in order to support accepting 'F'/'f' in SendDTMF.  This code does not enable sending Event 16 via dtmfmode=info.